+++ pothos-widgets/Periodogram/Periodogram.hpp
@@ -166,7 +166,7 @@ private slots:
     void installLegend(void);
     void handleLegendChecked(const QVariant &, bool, int);
     void handlePickerSelected(const QPointF &);
-    void handlePowerBins(const int index, const std::valarray<double> &bins);
+    void handlePowerBins(const int index, const std::valarray<float> &bins);
 
 private:
     QwtPlot *_mainPlot;
